requirements.txt
asgiref==3.4.1
boto3==1.18.40
botocore==1.21.40
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-storages==1.11.1
djangorestframework==3.12.4
idna==3.2
jmespath==0.10.0
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
